git: move credential helpers to exec path
If git-credential-osxkeychain isn't on your PATH, git will fail to find
it. Moving it to git's exec path[1] allows it to be located more
reliably. In other distributions such as Fedora[2] and Homebrew[3],
credential helpers are not added to the user's PATH unless they are
supplied by another package (e.g. git-credential-gcloud).

If there are concerns that people may be relying on referencing these
binaries directly we can add symlinks in bin for backwards
compatibility, but that's unlikely and easy to address.

[1]: https://git-scm.com/docs/git/2.46.0#Documentation/git.txt---exec-pathltpathgt
[2]: https://src.fedoraproject.org/rpms/git/blob/rawhide/f/git.spec#_651
[3]: https://github.com/Homebrew/homebrew-core/blob/12120ff79ce009489b958d8271af686d74403859/Formula/g/git.rb#L112